* simplify error checking logic by moving conditions within the function * move tests next to the file they test * re-word changelog message to emphasize that only temporary errors are retried
